requirements-noversions.txt
# torch>=1.8.1,<2
# torchvision>=0.8.9,<2
tqdm
scikit-learn
open_clip_torch
pycocoevalcap
webdataset
transformers
# new
typedparser
loguru
pandas
packg

